v1.5 Update. 2 new crates, and more Bug fixes
In the v1.5 update there are two new crates that want to join the
party. I have also cleaned up a bug that would let players pick up
Items if the plugin is unloaded.

Added:
- CrateOnTheGo, This is a physical crate you obtain by doing the give
command. It doesn’t matter if you put Virtual or Physical. You will
always get a physical crate that all you do is right click the ground
and it gives you a reward.
- FireCracker, This is a physical crate that is set with the /CC Create
command. It is not allowed as a virtual crate. A rocket is falling 25
blocks above the crate when opened. And when it lands the crate opens
up with a reward.

Bugs:
- When plugin is unloaded/Server Reload the items from QuickCrate are
left behind for players to pickup.