v1.5.7 Update
Hello there, in this update I have added a reward viewer to the crates.
If you left click a physical crate it now shows you all the rewards
that crate contins. For the virtual ones in the GUI all you have to do
is right click the crate and it will show you the rewards. I have also
added a Lore: option in the Crate.yml. Check the Default Crate.yml on
the main page and look under the rewards and you will see I added it
there. So if you want the lore to show up in the Reward viewer make
sure to add that Lore: Option. I have also added MaxRange: to the
rewards. You can add that also but if you don't it will auto be set to
100 untill you add it. What MaxRange is, is the max percent that the
chance is in. It is default at 100, so that means if the chance is 40
then it has a 40/100 chance of winning it. This is so if you want an
item to have a lower chance of getting picked then raise the MaxRange.

Added:
- Reward viewer for the crates.
- MaxRange to make items harder or easier to get.