v1.5.7.1 Update
Hello there, in this update it fixs a bug and an small error finder.
The error finder is for the rewards viewer. If there is an error in how
the items would load it will pop up with a Red Stained Clay block and
say Error and tell you the rewards that an error has happened at. It is
most likly a ID problem.

Bug:
- Added : support for the DisplayItem for the rewards.
Added:
- Small reward error finder in the Reward Viewer.