v1.5.7.3 Update
Hello, in this update I have made it so the plugin is more compatible
with all the 1.8 versions. I have also fixed a bug with QuadCrates. I
have added a new Timer to QuadCrates so that after a specified amount
of time it will kick them from the crate.

Bug:
- QuadCrate would teleport you to the crate when opening a physical
crate when there is no room.
- People could use ender pearls to teleport out of the Quad Crates
Added:
- Timer to kick players AFKing in Quad Crates
- Compatibility to all 1.8 versions