v1.5.9 Update || New Crate and Editable Names
Hello there, this update brings a new crate type to the plugin. Wonder
Crate is a GUI based crate. It has a very nice animation and as the
animation continues the intensity builds as you see less and less
prizes.

Changes:
- All default files if removed and now be easily regenerated by doing
/ce reload. You don't have to restart or reload any more to get a new
config.yml or any other missing files.
- GUI based crates can now have there names changed in the Crates file.
Added:
- The new Wonder Crate. Just go into your crate file and change crate
type to Wonder and you now have a wonder crate.
Bug:
- Items in the reward viewing GUIs sometimes stack. They now don't.