v1.6.5 Update || Performance Overhaul.
Hello, I have been working to overhaul the performance of this plugin.
I have been getting a bunch of servers that use around 80+ prizes in
one crate. The problem is because of GUI based crate types and from the
crate preview. This update should fix most lag that comes from these.

Bug Fixes:
  - Shift Right Clicking a physical crate with a physical key would
deny the player.
  - CSGO crate type now runs a bit smoother with less jerking.
  - Trying to use Quad Crate type as a virtual crate would deny you.
  - QuickCrate type would open as a quadcrate type when opening as a
virtual crate instead of denying you because the quickcrate type is
physical only.

Added:
  - Load the crate preview when the plugin loads and reloads.
  - Messages load in console when the crate preview is loaded. Example:
https://hastebin.com/lizegetapo.md
  - Update Checker option in the config to enable or disable the
checking of an update.
  - Lores now show up in the GUI based crate types.

Changes:
  - Big overhaul of the prize system. This should fix most lag problems
coming from the plugin.