v1.7 Update || 10th crate type.
Hello, this update contains the newest crate type which is called War.
I was requested to reproduce this by @zDan_YT on spigot. This is a
crate from the server called Treasure Wars. I hope you guys enjoy this
new crate type that I have added.

Here is the new crate type in action:
https://dl.dropboxusercontent.com/u/39364826/Pics/Crazy%20Crates/War.gif

Added:
  - A new crate type called War. This is a recreation of Treasure Wars
war chest. This was requested by @zDan_YT on spigot.
http://forum.treasurewars.net/index.php?threads/the-war-chest-update.104
7/
  - Error item for the display item if there is an error trying to show
it.

Bugs:
  - A debugging message was left in by mistake.

Chances:
  - More code clean up.