Fixed page size calculation.
There was a bug in how the amount of prizes that should be displayed per page was found. It did not take into account the boarder being on when having many prizes. Tested with 30 prizes and had preview size set to 3. This cause it to make only 3 pages of only 27 prizes and cut off the last 3.