Revert "Removed old checks of keys." reason below.
Reverting commit due to issues where users use normal items as keys. Example: if a crate is setup to use normal Nether Stars as a key for a crate it causes the plugin not to accept them. This method of keys are used as rewards for players that kill specific mobs to obtain their special items that they can then use to open a crate.

This reverts commit 15c0bf035d82d53e0a6719db7e33a651b6e09327.