Temp fix invalid permission message.
Just set the argument Boolean tabComplete in the permCheck method to true for now so as currently all it does it changes whether to notify the player or not.

I'm thinking that the arg should probably be changed to something more inclusive if used elsewhere to be something like notifyPlayer or something like that.