Update itembuilder
the ItemBuilder class now extends ParentItemBuilder which is in my library, Vital.

The overrides are necessary for the niche workflow setup at the moment.

I have a list of todo()'s labeled across the project and a various amount in my head. Do not touch or try to use this branch.

This branch is going to be the breaking changes of the breaking changes, All materials, potion effects, trim materials and patterns, enchantments and even how skulls work in the configurations will change.

All the ids will be completely different and a migration likely is just not going to happen ( I do not want that bloat everywhere, this is a hard break ), It's best to update when you feel like it or pick a time to make sure everything is up to date.