Updated /cc additem to be much more precise and friendly
Previously added rewards will still work until the next major version of Minecraft. The new /cc additem currently only supports 1 item at a time but it's an improvement from before. DisplayAmount is what tells the plugin how much of the item to give so that is required. It will default to 1!

If you add an `Items:` section, It will use that instead for rewards and treat the nbt tag as if it was just purely for display.

### Other changes
* Removed the config option `crate.unsupported-settings.old-key-checks` as a bug I fixed broke what this setting was used for.
* Updated how holograms are handled. FancyHolograms should be less finnicky and CMI should perform better.
* Remove holograms if the hologram isn't enabled just in case.
* Fixed an issue with materials not being recognized
* Used the already existing crate variable to play the sound.
* Added the ability to color the background using hex colors of holograms. It only works with CMI and FancyHolograms
 * You can check the `examples/crates` folder under `Hologram` section for an example of how to add colors.
 * `transparent` is an option that can be used as well for see through holograms which is the default.