Disable additem command temporarily
The additem command is being contested in it's usefulness, I also cannot decide on an approach to how to use the command. IT will be disabled temporarily until the rest of the plugin is in a state where adding the command back will not mess with anything.

This does mean that the previous tag DisplayNbt that was added will not work due to the methods used for that not existing, I cannot find the new method for it with recent changes in 1.20.6