add new tier preview menu
This framework used to build inventories, allows us to remove a lot of bloat into how it functions... including Tag checks. This allows for the execution to be a lot smoother, and right to the point due to how the items are handled internally. A lot of the fluff we had to manage is done automatically.