Add log to console/log to file back
It supports mass open, force open, etc.

It will zip up the logs folder on /crazycrates reload, plugin shutdown, plugin startup as to avoid any issues with files getting large.

Each zip file is time stamped, and numbered.

If a file is 0kb or empty, it will not be zipped.