Added ability to customize the csgo animation to either have the top/bottom be static or dynamic
If you run /crazycrates migrate -mt CratesDeprecated, The existing options you need will generate.

Please refer to the `examples/crates/CrateExample.yml` folder for a full example of how to do it.

It supports the same structure as the new way to do items in the `Items` section which can be found on the documentation.